+++ packages/nextjs/utils/abi-ninja/index.ts
@@ -1,0 +1,2 @@
+export * from "./chains";
+export * from "./proxyContracts";
